Added tests for create command

+++ src/Item.cpp
@@ -1,6 +1,6 @@
 #include "Item.hpp"
 
-Item::Item(std::string name, std::string seller, std::string bidder, int daysToAuction, int currentBid)
+Item::Item(std::string name, std::string seller, std::string bidder, int daysToAuction, double currentBid)
     : name(name), sellerUserName(seller), bidderUserName(bidder), daysToAuction(daysToAuction), currentBid(currentBid) {}
 
 const std::string &Item::GetName() const
@@ -23,7 +23,7 @@ const int &Item::GetDaysToAuction() const
     return daysToAuction;
 }
 
-const int &Item::GetCurrentBid() const
+const double &Item::GetCurrentBid() const
 {
     return currentBid;
 }
@@ -33,7 +33,7 @@ void Item::SetBidderName(std::string bidder)
     bidderUserName = bidder;
 }
 
-void Item::SetBid(int bid)
+void Item::SetBid(double bid)
 {
     currentBid = bid;
 }
